added support to remove all entities.

+++ engine/src/main/java/com/jukusoft/libgdx/rpg/engine/input/InputMapper.java
@@ -26,6 +26,6 @@ public interface InputMapper {
      *
      * @param keyCode keyCode of key on keyboard
     */
-    public int getActionCode (int keyCode);
+    public long getActionCode (int keyCode);
 
 }
